refactor: relocate CLI to a dedicated directory
