+++ src/cli/src/cmd.rs
